+++ src/main/scala/io/viash/ViashNamespace.scala
@@ -52,6 +52,9 @@ object ViashNamespace extends Logging {
         list.foreach(f)
   }
 
+  def targetOutputPath(targetDir: String, runnerId: String, config: Config): String =
+    targetOutputPath(targetDir, runnerId, config.namespace, config.name)
+
   def targetOutputPath(
     targetDir: String,
     runnerId: String,
@@ -88,7 +91,7 @@ object ViashNamespace extends Logging {
             if (flatten) {
               target
             } else {
-              targetOutputPath(target, runnerId, ns, funName)
+              targetOutputPath(target, runnerId, ac.config)
             }
           val nsStr = ns.map(" (" + _ + ")").getOrElse("")
           infoOut(s"Exporting $funName$nsStr =$runnerId=> $out")
@@ -127,7 +130,9 @@ object ViashNamespace extends Logging {
     tsv: Option[String] = None,
     append: Boolean = false,
     cpus: Option[Int],
-    memory: Option[String]
+    memory: Option[String],
+    dryRun: Option[Boolean] = None,
+    deterministicWorkingDirectory: Option[String] = None
   ): List[(AppliedConfig, ManyTestOutput)] = {
     val configs1 = configs.filter{tup => tup match {
       // remove nextflow because unit testing nextflow modules
@@ -149,7 +154,11 @@ object ViashNamespace extends Logging {
     val tsvExists = tsvPath.exists(Files.exists(_))
     val tsvWriter = tsvPath.map(Files.newBufferedWriter(_, StandardOpenOption.CREATE, StandardOpenOption.WRITE, StandardOpenOption.APPEND))
 
-    val parentTempPath = IO.makeTemp("viash_ns_test")
+    val parentTempPath = IO.makeTemp(
+      name = deterministicWorkingDirectory.getOrElse("viash_ns_test"),
+      parentTempPath = None,
+      addRandomized = deterministicWorkingDirectory.isEmpty
+    )
     if (keepFiles.getOrElse(true)) {
       info(s"The working directory for the namespace tests is ${parentTempPath.toString()}")
     }
@@ -173,6 +182,7 @@ object ViashNamespace extends Logging {
             val funName = ac.config.name
             val runnerName = ac.runner.get.id
             val engineName = ac.engines.head.id
+            val directoryName = if(namespace.isEmpty) funName else namespace.replace('/', '_') + "_" + funName
 
             // print start message
             infoOut(columnFormatString.format(namespace, funName, runnerName, engineName, "start", "", "", ""))
@@ -188,7 +198,9 @@ object ViashNamespace extends Logging {
                 quiet = true,
                 parentTempPath = Some(parentTempPath),
                 cpus = cpus,
-                memory = memory
+                memory = memory,
+                dryRun = dryRun,
+                deterministicWorkingDirectory = Some(directoryName)
               )
             } catch {
               case e: MissingResourceFileException =>