+++ src/main/scala/io/viash/cli/CLIConf.scala
@@ -252,14 +252,26 @@ class CLIConf(arguments: Seq[String]) extends ScallopConf(arguments) with Loggin
       "viash test",
       "Test the component using the tests defined in the viash config file.",
       "viash test config.vsh.yaml [-p docker] [-k true/false] [--setup cachedbuild]")
-    
+
     val setup = registerOpt[String](
       name = "setup",
       short = Some('s'),
       default = None,
       descr = "Which @[setup strategy](docker_setup_strategy) for creating the container to use [Docker Engine only]."
     )
-    
+    val dryRun = registerOpt[Boolean](
+      name= "dry_run",
+      default = Some(false),
+      descr = "Only generate the test script, do not run the test.",
+      hidden = true
+    )
+    val deterministicWorkingDirectory = registerOpt[String](
+      name = "deterministic_working_directory",
+      default = None,
+      descr = "Name of the working directory in the temporary directory. If not set, the working directory name will be a partially randomized name.",
+      hidden = true
+    )
+
     footer(
       s"""
          |The temporary directory can be altered by setting the VIASH_TEMP directory. Example:
@@ -349,6 +361,18 @@ class CLIConf(arguments: Seq[String]) extends ScallopConf(arguments) with Loggin
         short = Some('a'),
         default = Some(false),
         descr = "Append to tsv instead of overwrite"
+      )
+      val dryRun = registerOpt[Boolean](
+        name= "dry_run",
+        default = Some(false),
+        descr = "Only generate the test scripts, do not run them.",
+        hidden = true
+      )
+      val deterministicWorkingDirectory = registerOpt[String](
+        name = "deterministic_working_directory",
+        default = None,
+        descr = "Name of the working directory in the temporary directory. If not set, the working directory name will be a partially randomized name.",
+        hidden = true
       )
     }