+++ src/test/scala/io/viash/config/ConfigTest.scala
@@ -10,6 +10,8 @@ import io.circe.syntax._
 import io.viash.helpers.circe._
 import io.viash.helpers.data_structures._
 import io.viash.helpers.Logger
+import io.viash.engines.NativeEngine
+import io.viash.runners.ExecutableRunner
 
 class ConfigTest extends AnyFunSuite with BeforeAndAfterAll {
   Logger.UseColorOverride.value = Some(false)
@@ -45,6 +47,51 @@ class ConfigTest extends AnyFunSuite with BeforeAndAfterAll {
     
     val confParsed = conf.asJson.as[Config].fold(throw _, a => a)
     assert(confParsed == conf)
+  }
+
+  test("GetEngines and GetRunners with some engines and runners") {
+    val conf = Config(
+      name = "foo",
+      engines = List(
+        NativeEngine("engine0"),
+        NativeEngine("engine1"),
+        NativeEngine("engine2")
+      ),
+      runners = List(
+        ExecutableRunner("runner0"),
+        ExecutableRunner("runner1"),
+        ExecutableRunner("runner2")
+      )
+    )
+
+    assert(conf.findEngines(None) == conf.engines)
+    assert(conf.findRunners(None) == conf.runners)
+
+    assert(conf.findEngines(Some(".*")) == conf.engines)
+    assert(conf.findRunners(Some(".*")) == conf.runners)
+
+    assert(conf.findEngines(Some("engine0")) == List(NativeEngine("engine0")))
+    assert(conf.findEngines(Some("engine1")) == List(NativeEngine("engine1")))
+    assert(conf.findRunners(Some("runner0")) == List(ExecutableRunner("runner0")))
+    assert(conf.findRunners(Some("runner1")) == List(ExecutableRunner("runner1")))
+
+    assert(conf.findEngines(Some(".*0")) == List(NativeEngine("engine0")))
+    assert(conf.findEngines(Some(".*1")) == List(NativeEngine("engine1")))
+    assert(conf.findRunners(Some(".*0")) == List(ExecutableRunner("runner0")))
+    assert(conf.findRunners(Some(".*1")) == List(ExecutableRunner("runner1")))
+  }
+
+  test("GetEngines and GetRunners without engines or runners") {
+    val conf = Config(name = "foo")
+
+    assert(conf.findEngines(None) == List(NativeEngine("native")))
+    assert(conf.findRunners(None) == List(ExecutableRunner("executable")))
+
+    assert(conf.findEngines(Some(".*")) == List(NativeEngine("native")))
+    assert(conf.findRunners(Some(".*")) == List(ExecutableRunner("executable")))
+
+    assert(conf.findEngines(Some("native")) == List(NativeEngine("native")))
+    assert(conf.findRunners(Some("executable")) == List(ExecutableRunner("executable")))
   }
 
   // TODO: expand functionality tests